FIX issues raised during code review
